Interface methods from utils to opentelekomcloud
